// resources/ts/modules/book/type.ts
import type { Relationships } from "../page/type";
import type { Mangsach } from "../system/type";
import type { KieuBanQuyen } from "../topic/type";
import type { DonVi } from "../user/type";

export interface Sach {
    id: number;
    BanQuyen: boolean;
    BanQuyenTuNgay: string;
    BanQuyenDenNgay: string;
    BienTapBien: string;
    CanhBao: boolean;
    CoMSISBN: boolean;
    CreateBy: number;
    CreateOn: string;
    DaGui: boolean;
    Dai: string;
    DeCuong: string;
    DeTaiTuongTu: string;
    DiaChiCungCap: string;
    DiaChiTacGia: string;
    DiaChiTieuThu: string;
    DichGia: string;
    DinhDangTep: string;
    DungLuongTep: string;
    EditedBy: number;
    EditedOn: string;
    FMAVACH: string | null;
    GhiChu: string | null;
    GiaBia: number;
    HTXB: boolean;
    HoanThanh: boolean;
    ID_BoSach: number;
    ID_Cap: number;
    ID_DVLK: number;
    ID_DeTai: number;
    ID_DonVi: number;
    ID_LoaiXBP: number;
    ID_Lop: number;
    ID_MangSach: number;
    ID_MangSach_CXB: number;
    ID_MonHoc: number;
    ID_TuSach: number;
    ISBNCode: string;
    InUsed: boolean;
    IsCancel: boolean;
    IsDaDangKyLai: boolean | null;
    IsDeleted: boolean;
    IsSachDienTu: boolean;
    KhoaGuiNhan: string;
    KieuBanQuyen: KieuBanQuyen;
    LaDeTaiCKH: boolean | null;
    LaSachDich: boolean | null;
    LanTaiBan: number;
    LoaiChinhSua: number | null;
    LuaTuoi: string;
    LyDoThayDoiSoLuong: string | null;
    MaSo: string;
    MaSoCXB: string;
    MaSoGoc: string | null;
    MaSoQTG: string | null;
    MaVach: string | null;
    MauInBia: number;
    MauInRuot: number;
    NamTaiBan: string;
    NamXuatBan: string;
    NgayCapPhep: string;
    NgayDK: string;
    NgayCapQTG: string | null;
    NgayDocDuyet: string | null;
    NoiDung: string | null;
    PTXB: boolean | null;
    Rong: string;
    SoGPXB: string;
    SoHuuBanQuyen: string;
    SoLuong: number;
    SoLuongConLai: number;
    SoLuongTruocTamDung: number | null;
    SoTrang: number;
    TacGia: string;
    TenNguyenBan: string;
    TenSach: string;
    ThoiDiemCoDuBT: string;
    ThoiDiemRaSach: string;
    ThongTinBanQuyen: string;
    ThongTinLienQuan: string | null;
    ThongTinSachDich: string | null;
    TrangThaiDocBanThao: string | null;
    TypeLuaTuoi: number | null;
    VongThau: number | null;
    XetDuyetBanThao: boolean | null;
    YKienDocBanThao: string | null;

    don_vi: DonVi;
    mang_sach: Mangsach;
}

export interface FilterSach extends Relationships {
    title: string;
    ID_MangSach: number;
    ID_DonVi: number;
    NamXuatBan: string;
    NamTaiBan: string;
    HTXB: number;
    NgayDK: Date[];
    IsDeleted: boolean;
}

export interface PhieuChuyenBanThao {
    id: number;
    BienTapVien: string;
    CheBanCan: boolean;
    CoAoBoc: boolean;
    CreatedBy: number;
    CreatedOn: string;
    DaGui: boolean;
    Dai: number;
    DiaChiCungCap: string;
    DinhDangTep: string;
    DungLuongTep: string;
    EditedBy: number;
    EditedOn: string;
    GhiChu: string;
    ID_BTVNhan: number | null;
    ID_DV: number | null;
    ID_DeTai: number | null;
    ID_LanhDaoKiBenGui: number | null;
    ID_ListBienTapVien: string;
    ID_MangSach: number | null;
    ID_NguoiKy: number | null;
    ID_PhieuChuyenGoc: number | null;
    ID_Sach: number | null;
    InUsed: boolean;
    IsDeleted: boolean;
    IsSachDienTu: boolean;
    IsSubject: boolean;
    KhoSach: string;
    KhoaGuiNhan: string;
    LanIn: number;
    LoaiBia: boolean;
    LoaiPhieu: boolean;
    Locked: boolean;
    MaDVIN: string;
    MauInBia: number;
    MauInRout: number;
    NgayGiao: string;
    NgayNhan: string;
    NguoiGiao: string;
    NguoiNhan: string;
    Rong: number;
    SoBo: number;
    SoBoBanThao: number;
    SoBoBiaMau: number;
    SoBoPhimBia: number;
    SoMauInBia: number | null;
    SoTrang: number;
    SoTrangPhuBan: number;
    SoTrangRuotSach: number;
    TacGia: string;
}